test(manual): 第三轮穷尽排查，移出 tests/llm 下依赖真实 LLM API/外部数据的测试
CI 独立审查连续两轮抓到遗漏，本轮对整个 tests/ 目录（除 manual/performance）
做穷尽排查（grep config.jsonc/load_config/api_key/requests./socket./httpx.
逐个确认是否已被同文件 mock 覆盖）。新确认 8 个文件本质是手动集成测试：

- test_capswriter_segmentation.py / test_concurrent_calibration.py /
  test_llm_speaker_inference.py：真实 LLMClient + api_key，且依赖仓库外不
  存在的本地缓存转录文件（cache_dir\...\transcript_*.txt|json）
- test_chunk_validation.py / test_quality_validation_scoring.py：真实
  LLMClient 发起真实 API 调用（此前在无 config.jsonc 环境下会直接失败）
- test_summary_conditional_sections.py / test_summary_prompt_improvement.py：
  真实 LLMCoordinator + config.jsonc，后者甚至在配置缺失时 sys.exit(1)
- test_end_to_end_plain_text.py（本轮新查出，此前遗漏，因为它用
  api.context.get_config() 而非 load_config()，未命中上一轮的 grep 关键词）：
  真实 LLMCoordinator.process() 端到端调用 + 依赖仓库外的缓存转录文件

以上全部移入 tests/manual/，与已有 7 个文件的处理方式保持一致。

tests/unit/test_summary_fallback_config.py 拆分：其中 2 个用例断言的是真实
config.jsonc 的具体取值（部署配置检查，不是可测的产品逻辑），移入新增的
tests/manual/test_summary_fallback_config_real.py；断言 config.example.jsonc
结构和纯 LLMConfig.from_dict 解析的 3 个用例保留在 tests/unit/。

排查覆盖：tests/ 全目录 requests./httpx./socket. 用量逐文件复核 mock 覆盖，
sqlite3 连接均落在 tmp_path 临时库，sys.exit() 均在 __main__ 保护块内，
websocket/socket 测试均为本地 loopback 自建服务，均无需改动。

验证：删除/不存在 config/config.jsonc 环境下 `uv run --frozen pytest -q`
全绿（1262 passed，~23s，无 FAILED/ERROR，无非 ASCII 控制台输出）。

via [HAPI](https://hapi.run)
