+++ tests/manual/test_capswriter_segmentation.py
