+++ tests/manual/test_llm_speaker_inference.py
