+++ tests/manual/test_quality_validation_scoring.py
