+++ tests/manual/test_summary_conditional_sections.py
